# granja-corales-main/SolutionBackEnd/Solution.API/bin/Release/netcoreapp3.1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Solution.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:62718"
    }
  ],
  "paths": {
...
    "/api/Solicitudes": {
      "get": {
        "tags": [
          "Solicitudes"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitudes"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitudes"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitudes"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Solicitudes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              }
            }
          }
        }
      }
    },
    "/api/Solicitudes/{id}": {
      "get": {
        "tags": [
          "Solicitudes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Solicitudes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitudes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Solicitudes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitudes"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Solicitudes": {
        "type": "object",
        "properties": {
          "idSolicitud": {
            "type": "integer",
            "format": "int32"
          },
          "fecCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "estadoSolicitud": {
            "type": "string",
            "nullable": true
          },
          "idUsuario": {
            "type": "string",
            "nullable": true
          },
          "estadoAprobacion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}